[feature] created resource module
