+++ .gitignore
@@ -18,11 +18,13 @@ bucketlist.sqlite
 /bucketlist_api/custom_error.pyc
 /bucketlist_api/models.pyc
 /bucketlist_api/__init__.pyc
+/bucketlist_api/resources/__pycache__
 
 /tests/__pycache__
 /tests/test_api.pyc
 /tests/helpers.py
 /tests/test_models.pyc
+
 
 #cover
 /cover
